@@ -1,18 +1,19 @@
 module Api
   module V1
     class LocationPresenter < Brainstem::Presenter
-      presents "Location"
+      presents Location
 
       sort_order :id, "locations.id"
       sort_order :created_at, "locations.created_at"
 
       default_sort_order "id:desc"
 
-      def present(location)
-        {
-            :name           => location.name,
-            :widgets        => association(:widgets)
-        }
+      fields do
+        field :name, :string, "this Location's name"
+      end
+      
+      associations do
+        association :widgets, Widget, "the Widgets at this Location"
       end
     end
   end
